+++ cli/src/reporter.rs
@@ -30,26 +30,26 @@ arg_enum! {
 
 #[derive(Debug)]
 pub enum DumpAstError {
-    PGQueryError(PGQueryError),
-    IoError(std::io::Error),
-    JsonError(serde_json::error::Error),
+    PGQuery(PGQueryError),
+    Io(std::io::Error),
+    Json(serde_json::error::Error),
 }
 
 impl std::convert::From<PGQueryError> for DumpAstError {
     fn from(e: PGQueryError) -> Self {
-        Self::PGQueryError(e)
+        Self::PGQuery(e)
     }
 }
 
 impl std::convert::From<std::io::Error> for DumpAstError {
     fn from(e: std::io::Error) -> Self {
-        Self::IoError(e)
+        Self::Io(e)
     }
 }
 
 impl std::convert::From<serde_json::error::Error> for DumpAstError {
     fn from(e: serde_json::error::Error) -> Self {
-        Self::JsonError(e)
+        Self::Json(e)
     }
 }
 
@@ -57,7 +57,7 @@ pub fn dump_ast_for_paths<W: io::Write>(
     f: &mut W,
     paths: &[String],
     is_stdin: bool,
-    dump_ast: DumpAstOption,
+    dump_ast: &DumpAstOption,
 ) -> Result<(), DumpAstError> {
     let mut process_dump_ast = |sql: &str| -> Result<(), DumpAstError> {
         match dump_ast {
@@ -193,8 +193,7 @@ fn fmt_gcc<W: io::Write>(
                 .iter()
                 .map(|v| {
                     match v {
-                        ViolationMessage::Note(s) => s,
-                        ViolationMessage::Help(s) => s,
+                        ViolationMessage::Note(s) | ViolationMessage::Help(s) => s,
                     }
                     .to_string()
                 })
@@ -308,7 +307,7 @@ pub fn pretty_violations(
                 // 1-indexed
                 let lineno = sql[..start].lines().count() + 1;
 
-                let content = &sql[start..start + len + 1];
+                let content = &sql[start..=start + len];
 
                 // TODO(sbdchd): could remove the leading whitespace and comments to
                 // get cleaner reports
@@ -372,7 +371,7 @@ pub fn explain_rule<W: io::Write>(writer: &mut W, name: &str) -> Result<(), std:
     Ok(())
 }
 
-fn get_violations_emoji(count: usize) -> &'static str {
+const fn get_violations_emoji(count: usize) -> &'static str {
     if count > 0 {
         "🚒"
     } else {
@@ -741,7 +740,7 @@ SELECT 1;
         assert_debug_snapshot!(pretty_violations(violations, sql, filename));
     }
 
-    /// pretty_violations was slicing the SQL improperly, trimming off the first
+    /// `pretty_violations` was slicing the SQL improperly, trimming off the first
     /// letter.
     #[test]
     fn test_trimming_sql_newlines() {